Zeni dev (#8)
* start

* done hashing. TODO converstion from tree to string

* hasher completed

* add ecs - server connection

* add extra code

* extra code

* add code

* add extra comments

* KVMetadata and fixed build

* add extra code

* clean up repo

* finished merge and clean

* add ecs connection

* update code

* add server connection

* added update metadata

* fix bug

* built code

* add code

* add requests

* add code

* add code

* add code

* add code

* added metadata to kvstore

* add node

* add code

* succesful build

* added server hook for graceful exit

* extra changes

* code

* server start and shutdown properly

* extra cleanup

* fixed no data case

* some ui fixes

* added one new hash test

* finished hash tests, removed prints

* removed more print statements

* cleaned up the print statements

* added a few print statements

---------

+++ src/app_kvServer/Cache.java
@@ -1,31 +1,49 @@
 package app_kvServer;
 
 import java.util.*;
+import java.util.concurrent.locks.ReentrantReadWriteLock.WriteLock;
+
+import app_kvServer.exceptions.FailedException;
+import app_kvServer.exceptions.KeyNotFoundException;
+import app_kvServer.exceptions.WriteLockException;
+
 import java.io.*;
 
 import shared.messages.KVMessage.StatusType;
 
-//TODO real delete functionality
-//TODO move into KVServer
 class Cache extends Thread {
 
     LinkedHashMap<String, String> cache;
+    private boolean is_locked = false;
+    String dir;
 
-    public Cache(final int cacheSize) {
+    public Cache(final int cacheSize, String hostname, int port) {
         cache = new LinkedHashMap<String, String>() {
             @Override
             protected boolean removeEldestEntry(final Map.Entry eldest) {
                 return size() > cacheSize;
             }
         };
+        dir = String.format("./.cache/%s:%s", hostname, port);
+        createFolderIfNotExist(dir);
+    }
+
+    private void createFolderIfNotExist(String path) {
+        File directory = new File(path);
+        if (!directory.exists()) {
+            directory.mkdir();
+        }
     }
 
     private String getFilepath(String key) {
-        return String.format("./.cache/%s", key);
+        return String.format("%s/%s", dir, key);
     }
 
     // TODO MAKE THIS SYNCHRONIZED??
-    private void saveToDisk(String key, String value) {
+    private void saveToDisk(String key, String value) throws WriteLockException {
+        if (is_locked) {
+            throw new WriteLockException();
+        }
         String filepath = getFilepath(key);
         File file = new File(filepath);
         file.delete();
@@ -43,8 +61,27 @@ class Cache extends Thread {
         return file.exists();
     }
 
+    public LinkedHashMap<String, String> getAllKeyValues() {
+        File folder = new File(dir);
+        File[] listOfFiles = folder.listFiles();
+
+        if(listOfFiles == null){
+            return null;
+        }
+
+        LinkedHashMap<String, String> values =  new LinkedHashMap<String, String>();
+        for (File file : listOfFiles) {
+            try {
+                values.put(file.getName(), findFromDisk(file.getName()));
+            } catch (Exception e) {
+                System.out.println("Error getting all key values in Cache class");
+            }
+        }
+        return values;
+    }
+
     // TODO MAKE THIS SYNCHRONIZED??
-    private String findFromDisk(String key) {
+    private String findFromDisk(String key) throws FailedException {
         String filepath = getFilepath(key);
         File file = new File(filepath);
         try {
@@ -53,16 +90,27 @@ class Cache extends Thread {
             reader.close();
             return value;
         } catch (FileNotFoundException e) {
-            return " "; // TODO Change this to something better
+            throw new FailedException();
         }
     }
 
-    public StatusType save(String key, String value) {
+    public void setWriteLock(boolean locked) {
+        this.is_locked = locked;
+    }
+
+    public boolean getWriteLock() {
+        return this.is_locked;
+    }
+
+    public StatusType save(String key, String value) throws WriteLockException {
+        if (this.is_locked) {
+            throw new WriteLockException();
+        }
+
         StatusType status = StatusType.PUT;
         if (cache.containsKey(key)) {
             cache.remove(key);
             status = StatusType.PUT_UPDATE;
-
         }
         cache.put(key, value);
         saveToDisk(key, value);
@@ -70,26 +118,31 @@ class Cache extends Thread {
         return status;
     }
 
-    public String find(String key) throws Exception {
+    public String find(String key) throws KeyNotFoundException, FailedException {
         if (cache.containsKey(key)) {
-            System.out.println("Found in cache");
             return cache.get(key);
         }
         if (onDisk(key)) {
             return findFromDisk(key);
         }
-        throw new Exception("Could not find key");
+        throw new KeyNotFoundException();
     }
 
-    public void delete(String key) throws Exception {
-        if (!containsKey(key)) {
-            throw new Exception("Could not find key");
+    public void delete(String key, boolean forceDelete) throws KeyNotFoundException, WriteLockException {
+        if (this.is_locked && !forceDelete) {
+            throw new WriteLockException();
         }
+
+        if (!containsKey(key)) {
+            throw new KeyNotFoundException();
+        }
+
         if (cache.containsKey(key)) {
             cache.remove(key);
         }
 
         String filepath = getFilepath(key);
+
         File file = new File(filepath);
         if (file.exists()) {
             file.delete();
@@ -101,11 +154,17 @@ class Cache extends Thread {
         return cache.containsKey(key) || onDisk(key);
     }
 
-    public void clearCache() {
+    public void clearCache() throws WriteLockException {
+        if (this.is_locked) {
+            throw new WriteLockException();
+        }
         cache.clear();
     }
 
-    public void clearDisk() {
+    public void clearDisk() throws WriteLockException {
+        if (this.is_locked) {
+            throw new WriteLockException();
+        }
         File dir = new File(".cache");
         for (File file : dir.listFiles())
             if (!file.isDirectory())
@@ -115,20 +174,4 @@ class Cache extends Thread {
     public void printCache() {
         System.out.println(cache);
     }
-
-    // TODO: THIS IS TEST CODE?
-    // public void run() {
-    // save("1", "A");
-    // save("2", "B");
-    // save("3", "C");
-    // save("1", "D");
-    // save("4", "E");
-    // printCache();
-    // System.out.println(find("2"));
-    // }
-
-    // public static void main(String[] args) {
-    // new Cache(3).start();
-    // System.out.println("hello world");
-    // }
 }
